Relocate location tests to graphql crate
