+++ graphql/location/src/lib.rs
@@ -1,1 +1,348 @@
-pub mod mutations;
+mod mutations;
+use self::mutations::*;
+
+use async_graphql::*;
+use graphql_core::{
+    pagination::PaginationInput, standard_graphql_error::StandardGraphqlError, ContextExt,
+};
+use graphql_types::types::*;
+use repository::{LocationFilter, PaginationOption};
+
+#[derive(Default, Clone)]
+pub struct LocationQueries;
+
+#[Object]
+impl LocationQueries {
+    #[allow(non_snake_case)]
+    pub async fn apiVersion(&self) -> String {
+        "1.0".to_string()
+    }
+
+    /// Query omSupply "locations" entries
+    pub async fn locations(
+        &self,
+        ctx: &Context<'_>,
+        #[graphql(desc = "Pagination option (first and offset)")] page: Option<PaginationInput>,
+        #[graphql(desc = "Filter option")] filter: Option<LocationFilterInput>,
+        #[graphql(desc = "Sort options (only first sort input is evaluated for this endpoint)")]
+        sort: Option<Vec<LocationSortInput>>,
+    ) -> Result<LocationsResponse> {
+        let service_provider = ctx.service_provider();
+        let service_context = service_provider.context()?;
+
+        let locations = service_provider
+            .location_service
+            .get_locations(
+                &service_context,
+                page.map(PaginationOption::from),
+                filter.map(LocationFilter::from),
+                // Currently only one sort option is supported, use the first from the list.
+                sort.map(|mut sort_list| sort_list.pop())
+                    .flatten()
+                    .map(|sort| sort.to_domain()),
+            )
+            .map_err(StandardGraphqlError::from_list_error)?;
+
+        Ok(LocationsResponse::Response(LocationConnector::from_domain(
+            locations,
+        )))
+    }
+}
+
+#[derive(Default, Clone)]
+pub struct LocationMutations;
+
+#[Object]
+impl LocationMutations {
+    async fn insert_location(
+        &self,
+        ctx: &Context<'_>,
+        store_id: String,
+        input: InsertLocationInput,
+    ) -> InsertLocationResponse {
+        insert_location(ctx, &store_id, input)
+    }
+
+    async fn update_location(
+        &self,
+        ctx: &Context<'_>,
+        store_id: String,
+        input: UpdateLocationInput,
+    ) -> UpdateLocationResponse {
+        update_location(ctx, &store_id, input)
+    }
+
+    async fn delete_location(
+        &self,
+        ctx: &Context<'_>,
+        store_id: String,
+        input: DeleteLocationInput,
+    ) -> DeleteLocationResponse {
+        delete_location(ctx, &store_id, input)
+    }
+}
+
+#[cfg(test)]
+mod test {
+    use async_graphql::EmptyMutation;
+    use graphql_core::assert_graphql_query;
+    use graphql_core::test_helpers::setup_graphl_test;
+    use repository::{
+        mock::MockDataInserts, schema::LocationRow, Location, LocationFilter, LocationSort,
+        LocationSortField, StorageConnectionManager,
+    };
+    use repository::{EqualFilter, PaginationOption, Sort};
+    use serde_json::json;
+
+    use service::{
+        location::LocationServiceTrait,
+        service_provider::{ServiceContext, ServiceProvider},
+        ListError, ListResult,
+    };
+
+    use crate::LocationQueries;
+
+    type GetLocations = dyn Fn(
+            Option<PaginationOption>,
+            Option<LocationFilter>,
+            Option<LocationSort>,
+        ) -> Result<ListResult<Location>, ListError>
+        + Sync
+        + Send;
+
+    pub struct TestService(pub Box<GetLocations>);
+
+    impl LocationServiceTrait for TestService {
+        fn get_locations(
+            &self,
+            _: &ServiceContext,
+            pagination: Option<PaginationOption>,
+            filter: Option<LocationFilter>,
+            sort: Option<LocationSort>,
+        ) -> Result<ListResult<Location>, ListError> {
+            (self.0)(pagination, filter, sort)
+        }
+    }
+
+    pub fn service_provider(
+        location_service: TestService,
+        connection_manager: &StorageConnectionManager,
+    ) -> ServiceProvider {
+        let mut service_provider = ServiceProvider::new(connection_manager.clone());
+        service_provider.location_service = Box::new(location_service);
+        service_provider
+    }
+
+    #[actix_rt::test]
+    async fn test_graphql_locations_success() {
+        let (_, _, connection_manager, settings) = setup_graphl_test(
+            LocationQueries,
+            EmptyMutation,
+            "test_graphql_locations_success",
+            MockDataInserts::all(),
+        )
+        .await;
+
+        let query = r#"
+        query {
+            locations {
+              ... on LocationConnector {
+                nodes {
+                  id
+                  name
+                  code
+                  onHold
+                  stock {
+                      ... on StockLineConnector {
+                          nodes {
+                              id
+                          }
+                      }
+                  }
+                }
+                totalCount
+              }
+            }
+        }
+        "#;
+
+        // Test single record
+        let test_service = TestService(Box::new(|_, _, _| {
+            Ok(ListResult {
+                rows: vec![Location {
+                    location_row: LocationRow {
+                        id: "location_on_hold".to_owned(),
+                        name: "test_name".to_owned(),
+                        code: "test_code".to_owned(),
+                        on_hold: true,
+                        store_id: "store_a".to_owned(),
+                    },
+                }],
+                count: 1,
+            })
+        }));
+
+        let expected = json!({
+              "locations": {
+                  "nodes": [
+                      {
+                          "id": "location_on_hold",
+                          "name": "test_name",
+                          "code": "test_code",
+                          "onHold": true,
+                          "stock": {
+                              "nodes": [
+                                  {
+                                      "id": "stock_line_location_is_on_hold"
+                                  }
+                              ]
+                          }
+                      },
+                  ],
+                  "totalCount": 1
+              }
+          }
+        );
+
+        assert_graphql_query!(
+            &settings,
+            query,
+            &None,
+            &expected,
+            Some(service_provider(test_service, &connection_manager))
+        );
+
+        // Test no records
+
+        let test_service = TestService(Box::new(|_, _, _| {
+            Ok(ListResult {
+                rows: Vec::new(),
+                count: 0,
+            })
+        }));
+
+        let expected = json!({
+              "locations": {
+                  "nodes": [
+
+                  ],
+                  "totalCount": 0
+              }
+          }
+        );
+
+        assert_graphql_query!(
+            &settings,
+            query,
+            &None,
+            &expected,
+            Some(service_provider(test_service, &connection_manager))
+        );
+    }
+
+    #[actix_rt::test]
+    async fn test_graphql_locations_inputs() {
+        let (_, _, connection_manager, settings) = setup_graphl_test(
+            LocationQueries,
+            EmptyMutation,
+            "test_graphql_location_inputs",
+            MockDataInserts::all(),
+        )
+        .await;
+
+        let query = r#"
+        query(
+            $sort: [LocationSortInput]
+            $filter: LocationFilterInput
+          ) {
+            locations(sort: $sort, filter: $filter) {
+              __typename
+            }
+          }
+
+        "#;
+
+        let expected = json!({
+              "locations": {
+                  "__typename": "LocationConnector"
+              }
+          }
+        );
+
+        // Test sort by name no desc
+        let test_service = TestService(Box::new(|_, _, sort| {
+            assert_eq!(
+                sort,
+                Some(Sort {
+                    key: LocationSortField::Name,
+                    desc: None
+                })
+            );
+            Ok(ListResult::empty())
+        }));
+
+        let variables = json!({
+          "sort": [{
+            "key": "name",
+          }]
+        });
+
+        assert_graphql_query!(
+            &settings,
+            query,
+            &Some(variables),
+            &expected,
+            Some(service_provider(test_service, &connection_manager))
+        );
+
+        // Test sort by code with desc
+        let test_service = TestService(Box::new(|_, _, sort| {
+            assert_eq!(
+                sort,
+                Some(Sort {
+                    key: LocationSortField::Code,
+                    desc: Some(true)
+                })
+            );
+            Ok(ListResult::empty())
+        }));
+
+        let variables = json!({
+          "sort": [{
+            "key": "code",
+            "desc": true
+          }]
+        });
+
+        assert_graphql_query!(
+            &settings,
+            query,
+            &Some(variables),
+            &expected,
+            Some(service_provider(test_service, &connection_manager))
+        );
+
+        // Test filter
+        let test_service = TestService(Box::new(|_, filter, _| {
+            assert_eq!(
+                filter,
+                Some(LocationFilter::new().name(EqualFilter::equal_to("match_name")))
+            );
+            Ok(ListResult::empty())
+        }));
+
+        let variables = json!({
+          "filter": {
+            "name": { "equalTo": "match_name"},
+          }
+        });
+
+        assert_graphql_query!(
+            &settings,
+            query,
+            &Some(variables),
+            &expected,
+            Some(service_provider(test_service, &connection_manager))
+        );
+    }
+}
